Update Deck, Add HandDeck and Update Player
Update Deck: using Stacks instead of List, added method to draw multiple cards.

Add HandDeck: similar to deck but using ListBuffer and instead of has a method to draw multiple card, it has a method to add multiples cards.

Update Player: Implemented the method that use cards in hand and in deck (playCard and drawCard)

Deck and HandDeck is now stored in Deck package.

Added test for the new method of Deck and tests for the new methods of Player,